little change structure. Footer and Navbar components was moved to separated folder and has own style files
