+++ src/components/Navbar/Navbar.module.css
@@ -1,0 +1,17 @@
+.navbar {
+    width: 1270px;
+    height: 10px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 2rem;
+    background-color: #163A35;
+    border-radius: 20px;
+    margin-top: 30px;
+}
+.elips{
+    width: 45px;
+    height: 45px;
+    border-radius: 50%;
+    background-color: #10C298;
+}
